fix: add extra check exist files

+++ src/modules/fileManagement/index.js
@@ -6,10 +6,14 @@ import help from "../startManagement/index.js"
 
 const addFile = async (dirname, filename) => {
     if (await help.checkArg(filename.length, 1)) {
-        const PathTo = path.resolve(dirname, filename[0]);
-        await fsPromise.writeFile(PathTo, "", { flag: "wx" }, (err) => {
-            if (err) console.error("Operation failed");
-        });
+        try {
+            const PathTo = path.resolve(dirname, filename[0]);
+            await fsPromise.writeFile(PathTo, "", { flag: "wx" }, (err) => {
+                if (err) console.error("Operation failed");
+            });
+        } catch (err) {
+            console.error("Operation failed");
+        }
     }
 };
 
@@ -17,12 +21,15 @@ const readFile = async (dirname, filename) => {
     if (await help.checkArg(filename.length, 1)) {
         const pathToRead = path.resolve(dirname, filename[0]);
         try {
-            const content = fs.createReadStream(pathToRead, { encoding: 'utf-8' });
-            await pipeline(content, process.stdout, { end: false });
+            const stats = await fsPromise.stat(pathToRead);
+            if (stats.isFile()) {
+                const content = fs.createReadStream(pathToRead, { encoding: 'utf-8' });
+                await pipeline(content, process.stdout, { end: false });
 
-            content.on('error', (err) => {
-                console.error("Operation failed");
-            });
+                content.on('error', (err) => {
+                    console.error("Operation failed");
+                });
+            }
         } catch (err) {
             console.error("Operation failed");
         }
@@ -33,11 +40,18 @@ const renameFile = async (dirname, files) => {
     if (await help.checkArg(files.length, 2)) {
         const pathOld = path.resolve(dirname, files[0]);
         const filename = path.basename(files[1]);
-        const pathNew = path.resolve(dirname, files[0], "..", filename);
-        try {
-            await fsPromise.rename(pathOld, pathNew);
-        } catch {
-            console.log('Operation failed')
+        if (filename == files[1]) {
+            const pathNew = path.resolve(dirname, files[0], "..", filename);
+            try {
+                const stats = await fsPromise.stat(pathOld);
+                if (stats.isFile()) {
+                    await fsPromise.rename(pathOld, pathNew);
+                }
+            } catch {
+                console.log('Operation failed')
+            }
+        } else {
+            console.log('Invalid input')
         }
     }
 };
@@ -49,9 +63,12 @@ const copyFile = async (dirname, files) => {
         const pathTo = path.resolve(dirname, files[1], filename);
 
         try {
-            const copyFrom = fs.createReadStream(pathFrom, "utf-8");
-            const copyTo = fs.createWriteStream(pathTo, "utf-8");
-            await pipeline(copyFrom, copyTo);
+            const stats = await fsPromise.stat(pathFrom);
+            if (stats.isFile()) {
+                const copyFrom = fs.createReadStream(pathFrom, "utf-8");
+                const copyTo = fs.createWriteStream(pathTo, "utf-8");
+                await pipeline(copyFrom, copyTo);
+            }
         } catch (err) {
             console.error("Operation failed");
         }
@@ -62,7 +79,10 @@ const removeFile = async (dirname, filename) => {
     if (await help.checkArg(filename.length, 1)) {
         try {
             const pathToDelete = path.join(dirname, filename[0]);
-            await fsPromise.unlink(pathToDelete);
+            const stats = await fsPromise.stat(pathToDelete);
+            if (stats.isFile()) {
+                await fsPromise.unlink(pathToDelete);
+            }
         } catch (err) {
             console.error("Operation failed");
         }
@@ -71,10 +91,19 @@ const removeFile = async (dirname, filename) => {
 
 const moveFile = async (dirname, files) => {
     if (await help.checkArg(files.length, 2)) {
+        const pathFrom = path.resolve(dirname, files[0]);
+        const filename = path.basename(files[0]);
+        const pathTo = path.resolve(dirname, files[1], filename);
+
         try {
-            await copyFile(dirname, files);
-            await removeFile(dirname, files.slice(0, 1));
-        } catch {
+            const stats = await fsPromise.stat(pathFrom);
+            if (stats.isFile()) {
+                const copyFrom = fs.createReadStream(pathFrom, "utf-8");
+                const copyTo = fs.createWriteStream(pathTo, "utf-8");
+                await pipeline(copyFrom, copyTo);
+                await fsPromise.unlink(pathFrom);
+            }
+        } catch (err) {
             console.error("Operation failed");
         }
     }
